requirements.txt
discord.py
loguru
nest-asyncio
pandas
pybit
python-dotenv
ta
